Using quoted macro
