+++ airframe-log/shared/src/main/scala-2.x/wvlet/log/LoggerBase.scala
@@ -37,3 +37,31 @@ trait LoggerBase {
   def trace(message: Any, cause: Throwable): Unit =
     macro traceLogMethodWithCause
 }
+
+/**
+  */
+trait LoggingMethods extends Serializable {
+  import wvlet.log.LogMacros._
+
+  protected def error(message: Any): Unit = macro errorLog
+  protected def error(message: Any, cause: Throwable): Unit =
+    macro errorLogWithCause
+
+  protected def warn(message: Any): Unit = macro warnLog
+  protected def warn(message: Any, cause: Throwable): Unit =
+    macro warnLogWithCause
+
+  protected def info(message: Any): Unit = macro infoLog
+  protected def info(message: Any, cause: Throwable): Unit =
+    macro infoLogWithCause
+
+  protected def debug(message: Any): Unit = macro debugLog
+  protected def debug(message: Any, cause: Throwable): Unit =
+    macro debugLogWithCause
+
+  protected def trace(message: Any): Unit = macro traceLog
+  protected def trace(message: Any, cause: Throwable): Unit =
+    macro traceLogWithCause
+
+  protected def logAt(logLevel: LogLevel, message: Any): Unit = macro logAtImpl
+}
